Reorganize file structure, ssr -> server
