+++ src/client/index.js
@@ -1,9 +1,10 @@
+/* eslint-disable no-param-reassign, guard-for-in, no-restricted-syntax */
+
 import createDomElem from './createDomElem'
-import elements from '../domElements'
+import elements from './domElements'
 import checkAttr from './checkAttr'
 
 const domProps = elements.reduce((acc, elem: string) => {
-  /* eslint-disable no-param-reassign, guard-for-in, no-restricted-syntax */
   for (const key in Object.getPrototypeOf(createDomElem(elem))) acc[key] = true
 
   return acc
