+++ src/client/reactProps.js
@@ -1,8 +1,9 @@
-import * as whitelists from '../whitelists'
+/* eslint-disable guard-for-in, no-restricted-syntax */
+
+import * as whitelists from './whitelists'
 
 const reactProps = {}
 
-/* eslint-disable guard-for-in, no-restricted-syntax */
 for (const key in whitelists) Object.assign(reactProps, whitelists[key])
 
 export default reactProps
