+++ src/client/whitelists/svg.js
