+++ src/server/index.js
@@ -1,8 +1,9 @@
+/* eslint-disable no-param-reassign, no-restricted-syntax */
+
 import attrs from './htmlAttributes'
 import checkAttr from './checkAttr'
 
 const domProps = Object.keys(attrs).reduce((acc, elem: string) => {
-  /* eslint-disable no-param-reassign, no-restricted-syntax */
   for (const attr of attrs[elem]) acc[attr] = true
 
   return acc
