+++ src/tests/server/validAttr.test.js
@@ -1,5 +1,5 @@
 import validAttrTests from '../validAttrTests'
-import validAttr from '../../ssr'
+import validAttr from '../../server'
 
 describe('validAttr for server', () => {
   validAttrTests(validAttr)
